+++ projects/rocshmem/tests/functional_tests/verify_results_kernels.hpp
@@ -30,7 +30,7 @@ namespace rocshmem {
 [[maybe_unused]] static __global__ void verify_results_kernel_char(char *source, char *dest, size_t buf_size,
                                                   bool *verification_error) {
 
-  int idx = get_flat_id();
+  size_t idx = get_flat_id();
 
   if (idx >= buf_size) {
     return;
